FIX work together npm link and npm install
