+++ index.js
@@ -1,2 +1,2 @@
 require('babel-core/register');
-module.exports = require('./lib').default;
+module.exports = require('./lib/api').default;
